abstraction for shooting towers
