+++ Assets/Scripts/Towers/Archer.cs
@@ -2,37 +2,9 @@ using System.Collections;
 using System.Linq;
 using UnityEngine;
 
-public class Archer : MonoBehaviour
+public class Archer : BaseTower
 {
-    public float range = 4;
-    public float cooldown = 1;
-    public float damage = 10;
-
-    public TowerHelpers.TowerTargetTypes targetType = TowerHelpers.TowerTargetTypes.CLOSEST_TO_FINISH;
-
-    private bool canShoot = true;
-
-
-    void FixedUpdate()
-    {
-        if (!canShoot) return;
-
-        GameObject[] enemies = TowerHelpers.GetEnemiesInRange(transform.position, range);
-        if (enemies.Length == 0) return;
-
-        GameObject target = TowerHelpers.SelectEnemyToAttack(enemies, targetType);
-        StartCoroutine(AnimateArrow(target));
-        canShoot = false;
-        StartCoroutine(ResetCooldown());
-    }
-
-    IEnumerator ResetCooldown()
-    {
-        yield return new WaitForSeconds(cooldown);
-        canShoot = true;
-    }
-
-    IEnumerator AnimateArrow(GameObject enemy)
+    protected override IEnumerator AnimateProjectile(GameObject enemy)
     {
         GameObject arrow = GameObject.CreatePrimitive(PrimitiveType.Cylinder);
         arrow.transform.SetParent(transform);
@@ -40,12 +12,12 @@ public class Archer : MonoBehaviour
         arrow.transform.localScale = new Vector3(0.1f, 0.1f, 0.1f);
         arrow.GetComponent<Renderer>().material.color = Color.red;
 
-        yield return TowerHelpers.AnimateBezierProjectile(arrow, transform.position, enemy, 2, 1, KillArrow, TowerHelpers.TowerProjectileRotationTypes.LOOK_AT_TARGET);
+        yield return TowerHelpers.AnimateBezierProjectile(arrow, transform.position, enemy, 2, 1, KillProjectile, TowerHelpers.TowerProjectileRotationTypes.LOOK_AT_TARGET);
     }
 
-    public void KillArrow(GameObject arrow, GameObject enemy, Vector3 _enemyPosition)
+    protected override void KillProjectile(GameObject projectile, GameObject enemy, Vector3 _enemyPosition)
     {
-        Destroy(arrow);
-        if (enemy != null) enemy.GetComponent<BaseEnemy>().TakeDamage((int)damage, DamageTypes.PHYSICAL);
+        Destroy(projectile);
+        enemy?.GetComponent<BaseEnemy>().TakeDamage((int)damage, DamageTypes.PHYSICAL);
     }
 }
