+++ Assets/Scripts/Towers/BaseTower.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 261ee23baf38c8e44afdc7e83279e157
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
