+++ Assets/Scripts/Towers/Magic.cs
@@ -1,38 +1,9 @@
 using System.Collections;
-using System.Linq;
 using UnityEngine;
 
-public class Magic : MonoBehaviour
+public class Magic : BaseTower
 {
-
-    public float range = 4;
-    public float cooldown = 1;
-    public float damage = 10;
-
-    public TowerHelpers.TowerTargetTypes targetType = TowerHelpers.TowerTargetTypes.CLOSEST_TO_FINISH;
-
-    private bool canShoot = true;
-
-    void FixedUpdate()
-    {
-        if (!canShoot) return;
-
-        GameObject[] enemies = TowerHelpers.GetEnemiesInRange(transform.position, range);
-        if (enemies.Length == 0) return;
-
-        GameObject target = TowerHelpers.SelectEnemyToAttack(enemies, targetType);
-        StartCoroutine(AnimateSphere(target));
-        canShoot = false;
-        StartCoroutine(ResetCooldown());
-    }
-
-    IEnumerator ResetCooldown()
-    {
-        yield return new WaitForSeconds(cooldown);
-        canShoot = true;
-    }
-
-    IEnumerator AnimateSphere(GameObject enemy)
+    protected override IEnumerator AnimateProjectile(GameObject enemy)
     {
         GameObject sphere = GameObject.CreatePrimitive(PrimitiveType.Sphere);
         sphere.transform.SetParent(transform);
@@ -40,10 +11,10 @@ public class Magic : MonoBehaviour
         sphere.transform.localScale = new Vector3(0.1f, 0.1f, 0.1f);
         sphere.GetComponent<Renderer>().material.color = Color.blue;
 
-        yield return TowerHelpers.AnimateDirectProjectile(sphere, enemy, 5, KillSphere);
+        yield return TowerHelpers.AnimateDirectProjectile(sphere, enemy, 5, KillProjectile);
     }
 
-    void KillSphere(GameObject sphere, GameObject enemy, Vector3 _enemyPosition)
+    protected override void KillProjectile(GameObject sphere, GameObject enemy, Vector3 _enemyPosition)
     {
         Destroy(sphere);
         if (enemy != null) enemy.GetComponent<BaseEnemy>().TakeDamage((int)damage, DamageTypes.MAGIC);
